requirements.txt
langchain==0.3.25
langchain_community==0.3.24
langchain_mistralai==0.2.10
langchain_openai==0.3.18
python-dotenv==1.1.0
streamlit==1.42.0
